Refactor sct_header into sct_image option
This is more intuitive for users. Also, designing an sct_header
interface should be done alongside other header utilities, not
in speculation.
