+++ setup.py
@@ -94,7 +94,6 @@ setup(
                 'sct_fmri_compute_tsnr',
                 'sct_fmri_moco',
                 'sct_get_centerline',
-                'sct_header',
                 'sct_image',
                 'sct_label_utils',
                 'sct_label_vertebrae',
