+++ spinalcordtoolbox/image.py
@@ -17,11 +17,13 @@ import itertools
 import warnings
 import logging
 import shutil
+import math
 from typing import Sequence
 
 import nibabel as nib
 import numpy as np
 import pathlib
+from contrib import fslhd
 
 import transforms3d.affines as affines
 import re
@@ -1558,3 +1560,77 @@ def pad_image(im: Image, pad_x_i: int = 0, pad_x_f: int = 0, pad_y_i: int = 0, p
     im_out.hdr.structarr['srow_z'][-1] = new_origin[2]
 
     return im_out
+
+
+HEADER_FORMATS = ('sct', 'fslhd', 'nibabel')
+
+
+def create_formatted_header_string(header, output_format='sct'):
+    """
+    Generate a string with formatted header fields for pretty-printing.
+
+    :param header: Input header to apply formatting to.
+    :param output_format: Specify how to format the output header.
+    """
+    if output_format == 'sct':
+        formatted_fields = _apply_sct_header_formatting(fslhd.generate_nifti_fields(header))
+        aligned_string = _align_dict(formatted_fields)
+    elif output_format == 'fslhd':
+        formatted_fields = fslhd.generate_nifti_fields(header)
+        aligned_string = _align_dict(formatted_fields)
+    elif output_format == 'nibabel':
+        formatted_fields = {k: v[()] for k, v in dict(header).items()}
+        aligned_string = _align_dict(formatted_fields, use_tabs=False, delimiter=": ")
+    else:
+        raise ValueError(f"Can't format header using '{output_format}' format. Available formats: {HEADER_FORMATS}")
+
+    return aligned_string
+
+
+def _apply_sct_header_formatting(fslhd_fields):
+    """
+    Tweak fslhd's header fields using SCT's visual preferences.
+
+    :param fslhd_fields: Dict with fslhd's header fields.
+    :return modified_fields: Dict with modified header fields.
+    """
+    modified_fields = {}
+    dim, pixdim = [], []
+    for key, value in fslhd_fields.items():
+        # Replace split dim fields with one-line dim field
+        if key.startswith('dim'):
+            dim.append(value)
+            if key == 'dim7':
+                modified_fields['dim'] = dim
+        # Replace split pixdim fields with one-line pixdim field
+        elif key.startswith('pixdim'):
+            pixdim.append(float(value))
+            if key == 'pixdim7':
+                modified_fields['pixdim'] = pixdim
+        # Leave all other fields
+        else:
+            modified_fields[key] = value
+
+    return modified_fields
+
+
+def _align_dict(dictionary, use_tabs=True, delimiter=""):
+    """
+    Create a string with aligned padding from a dict's keys and values.
+
+    :param dictionary: Variable of type dict.
+    :param use_tabs: Whether to use tabs instead of spaces for padding.
+
+    :return: String containing padded dict key/values.
+    """
+    len_max = max([len(str(name)) for name in dictionary.keys()]) + 2
+    out = []
+    for k, v in dictionary.items():
+        if use_tabs:
+            len_max = int(8 * round(float(len_max)/8))  # Round up to the nearest 8 to align with tab stops
+            padding = "\t" * math.ceil((len_max - len(k))/8)
+        else:
+            padding = " " * (len_max - len(k))
+        out.append(f"{k}{padding}{delimiter}{v}")
+    return '\n'.join(out)
+
